fixes aggregations
previously we selected the 'model_run' column, but the regex also matched 'model_runs'.
in v1.2.0 we started to keep the full model_runs for all aggregatsions, so the grouping
broke.

this change specifically targets the 'model_run' column (aswell as the year column),
which only exist when the 'model_runs' or 'time_profiles' columns are unnested.